+++ src/scannertwain.h
@@ -14,6 +14,9 @@ public:
     ~ScannerTwain();
 
 public:
+    bool Scan(QString filename);
+
+private:
     bool LoadDLL();
     bool UnLoadDLL();
     bool OpenSourceManager();
@@ -22,11 +25,9 @@ public:
     bool CloseSource();
     bool Enabled();
     bool Disabled();
-    bool TransferImage();
-    bool WriteImage(LPBITMAPINFOHEADER	bi);
+    bool TransferImage(QString filename);
+    bool WriteImage(LPBITMAPINFOHEADER	bi, QString filename);
     bool TwainSetting(void);
-
-private:
     float    Fix32ToFloat (TW_FIX32 fix32);
     TW_FIX32 FloatToFix32 (float floater);